Add files via upload
Here is the internship screening task:
Concept Statement:

Develop a KBC-style game where the question and a QR code are displayed on a computer screen. Players who wish to participate scan the QR code using their mobile devices, enter their names, and view the same question on their screen. If a player provides the correct answer, a "Congratulations" message, along with the player's name, should appear on the computer screen, which will then move to the next question. If the answer is incorrect, the player will receive a message on their mobile device stating the answer is wrong.

Include 5 sample questions.
You are free to use the framework/language of your choice.
Estimated work hours: 8.